add applePayPaymentToken to PaymentRequest - rename MiscellaneousHandler to WalletHandler - deprecate requestApplePaySession(String validationUrl, String domain) in favor of requestApplePaySession(ApplePaySessionRequest body)
